rename some files and clean up general structure
